♻️ Refactor : 강아지 프로필 component 분리
