+++ src/pages/ProfilePage/index.tsx
@@ -4,6 +4,7 @@ import { Separator } from '~components/Separator'
 import { Typo13, Typo15, Typo20, Typo24 } from '~components/Typo'
 import * as S from './styles'
 
+import DogProfile from '~components/DogProfile'
 export default function ProfilePage() {
   const { id } = useParams()
   const navigate = useNavigate()
@@ -40,35 +41,7 @@ export default function ProfilePage() {
         </S.WalkInfoWrapper>
       </S.WalkInfoArea>
 
-      <S.DogInfoArea>
-        <S.DogInfoWrapper>
-          <Profile $size={80} $src={dogInfo.profileImg} />
-          <S.DogDetailWrapper>
-            <S.TypoWrapper>
-              <Typo20 $weight='700'>{dogInfo.name}</Typo20>
-              <Typo15 $weight='400'>{dogInfo.breed}</Typo15>
-              <Separator $height={8} />
-              <Typo15 $weight='400'>{dogInfo.age}살</Typo15>
-              <Separator $height={8} />
-              <Typo15 $weight='400'>{dogInfo.gender === 'male' ? '남' : '여'}</Typo15>
-            </S.TypoWrapper>
-            <S.TypoWrapper>
-              <Typo13>중성화 X</Typo13>
-              <Separator $height={8} />
-              <Typo13>3.4KG</Typo13>
-            </S.TypoWrapper>
-          </S.DogDetailWrapper>
-        </S.DogInfoWrapper>
-
-        <S.OneLineIntro>
-          <Typo15 $weight='800' $color='default'>
-            우리 댕댕이를 소개해요!
-          </Typo15>
-          <Typo13 $weight='500' style={{ lineHeight: 1.2 }}>
-            {dogInfo.intro}
-          </Typo13>
-        </S.OneLineIntro>
-      </S.DogInfoArea>
+      <DogProfile />
     </S.ProfilePage>
   )
 }
@@ -82,16 +55,4 @@ const userInfo = {
   walkDistance: 32,
   gangbunttaCount: 16,
   profileImg: '',
-}
-
-const dogInfo = {
-  name: '밤톨이',
-  breed: '포메라니안',
-  age: 4,
-  gender: 'male',
-  neutered: false, // 중성화 여부
-  weight: 3.4,
-  profileImg: '',
-  intro: `우리아이 안 물어요 착해요. 
-강아지껌을 너무 좋아해요 같이 놀아요. `,
 }
